pyproject.toml
[project]
name = "connectdeaf-ia-faq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-formrecognizer==3.3.3",
    "azure-identity==1.18.0",
    "azure-search-documents==11.5.2",
    "azure-storage-blob==12.24.1",
    "fastapi==0.115.6",
    "install",
    "langchain==0.3.9",
    "openai==1.59.6",
    "pandas==2.2.3",
    "pip",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pydantic-settings==2.6.1",
    "uvicorn==0.32.1",
]
